requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.5.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.4.1
cryptography==41.0.3
Django==4.0.3
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.2.1
exceptiongroup==1.1.3
fastapi==0.103.1
Flask==2.3.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
interchange==2021.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
monotonic==1.6
neo4j==5.11.0
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.1.1
packaging==23.1
pansi==2020.7.3
Pillow==10.0.0
psycopg2-binary==2.9.7
py2neo==2021.2.3
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.15.1
PyJWT==2.8.0
pykerberos==1.2.4
pymongo==3.12.1
pyspnego==0.9.2
python-dotenv==1.0.0
python-snappy==0.6.1
pytz==2022.1
pywinrm==0.4.3
requests==2.31.0
requests-ntlm==1.2.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.2.4
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==2.3.6
xmltodict==0.13.0
